pyproject.toml
[project]
name = "neo-assetmaker"
version = "2.4.0"
description = "明日方舟通行证素材制作器"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyQt6>=6.5.0",
    "PyQt6-WebEngine>=6.5.0",
    "PyQt6-Fluent-Widgets>=1.5.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "jsonschema>=4.17.0",
    "thefuzz>=0.22.0",
    "python-Levenshtein>=0.25.0",
    "requests>=2.32.0",
    # 扩展模块依赖
    "qtpy>=2.4",
    "httpx>=0.27",
    "keyring>=25",
    "platformdirs>=4",
    "pyusb>=1.2",
    "fido2>=2.1",
]

[dependency-groups]
dev = [
    "cx_Freeze>=6.15.0",
]
